Update dependabot config to include security updates only
The Dependabot configuration in GitHub workflows has been modified. The previous config ran a dry run of all updates. The change makes it so that the script now runs a dry run of only the security updates. This ensures that any executed updates are pertinent to the security of the repository, reducing noise from non-essential updates.