test: Update token and repository references in workflow files
The commit updates the `.github/workflows/test-scan.yaml` and `.github/workflows/dependabot-scan.yaml` workflow files. It replaces the `secrets.ACCESS_TOKEN` with `secrets.GITHUB_TOKEN` and introduces a new input parameter 'repository' to replace the older reference. This change is aimed to ensure compatibility with the current Github secrets and repository usage policy.